+++ example/lib/screens/home/home_screen.dart
@@ -1,0 +1,15 @@
+// ignore_for_file: unused_import
+import 'dart:async';
+
+import 'package:clean_archt_lint_example/core/app_exemple.dart';
+import 'package:clean_archt_lint_example/data/models/product_model.dart';
+import 'package:flutter/widgets.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}